backend/requirements.txt
# requirements.txt

# Core dependencies
# Let google-adk determine the correct versions of fastapi and pydantic
fastapi
pydantic
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
httpx>=0.26.0

# Auth dependencies
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose[cryptography]==3.3.0

# Google ADK (includes AI capabilities)
google-adk>=1.17.0
google-generativeai

# Extra for form data
python-multipart
ag_ui_adk